Consistent Hashing Key Long Type으로 변경
work items: #61

+++ key-value-managed-proxy-server/src/main/kotlin/com/tommy/proxy/consistenthashing/hash/HashFunction.kt
@@ -2,5 +2,5 @@ package com.tommy.proxy.consistenthashing.hash
 
 interface HashFunction {
 
-    fun doHash(key: String): Int
+    fun doHash(key: String): Long
 }
